+++ non-streaming/SQL/rendered_frame_post_processed_table.sql
@@ -1,5 +1,4 @@
-CREATE TABLE doomhouse.rendered_frame_1
-(
+CREATE TABLE doomhouse_ns.rendered_frame_post_processed_1 (
     pos_x Float32,
     pos_y Float32,
     image_data Array(UInt32)
@@ -7,8 +6,7 @@ CREATE TABLE doomhouse.rendered_frame_1
 ENGINE = Memory 
 SETTINGS min_rows_to_keep = 1, max_rows_to_keep = 1;
 
-CREATE TABLE doomhouse.rendered_frame_2
-(
+CREATE TABLE doomhouse_ns.rendered_frame_post_processed_2 (
     pos_x Float32,
     pos_y Float32,
     image_data Array(UInt32)
@@ -16,8 +14,7 @@ CREATE TABLE doomhouse.rendered_frame_2
 ENGINE = Memory 
 SETTINGS min_rows_to_keep = 1, max_rows_to_keep = 1;
 
-CREATE TABLE doomhouse.rendered_frame_3
-(
+CREATE TABLE doomhouse_ns.rendered_frame_post_processed_3 (
     pos_x Float32,
     pos_y Float32,
     image_data Array(UInt32)
@@ -25,8 +22,7 @@ CREATE TABLE doomhouse.rendered_frame_3
 ENGINE = Memory 
 SETTINGS min_rows_to_keep = 1, max_rows_to_keep = 1;
 
-CREATE TABLE doomhouse.rendered_frame_4
-(
+CREATE TABLE doomhouse_ns.rendered_frame_post_processed_4 (
     pos_x Float32,
     pos_y Float32,
     image_data Array(UInt32)